# ahms-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ayurvedic Hospital Management System — Phase 1 API (Auth, Roles, Departments, Patients, Doctors, Appointments, Dashboard).",
        "title": "AHMS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/medicines": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pharmacy"
                ],
                "summary": "List medicines (optionally low-stock only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by name/formulation",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only medicines at or below threshold",
                        "name": "low_stock",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/pharmacy.MedicineResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pharmacy"
                ],
                "summary": "Create a medicine (stock master)",
                "parameters": [
                    {
                        "description": "Medicine",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pharmacy.CreateMedicineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pharmacy.MedicineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/medicines/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pharmacy"
                ],
                "summary": "Get a medicine by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pharmacy.MedicineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pharmacy"
                ],
                "summary": "Update a medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicine",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pharmacy.UpdateMedicineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pharmacy.MedicineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/medicines/{id}/stock": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pharmacy"
                ],
                "summary": "Adjust medicine stock (purchase / adjustment)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stock adjustment",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pharmacy.StockAdjustRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pharmacy.MedicineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/prescriptions/{id}/dispense": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records actual dispensing, decrements stock, updates prescription status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pharmacy"
                ],
                "summary": "Dispense medicines against a prescription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prescription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dispense details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pharmacy.DispenseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pharmacy.PrescriptionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pharmacy.CreateMedicineRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "batch_number": {
                    "type": "string",
                    "maxLength": 100
                },
                "expiry_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "formulation": {
                    "type": "string",
                    "maxLength": 100
                },
                "low_stock_threshold": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "stock_qty": {
                    "type": "number",
                    "minimum": 0
                },
                "unit": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "pharmacy.DispenseItem": {
            "type": "object",
            "required": [
                "prescription_item_id",
                "quantity"
            ],
            "properties": {
                "medicine_id": {
                    "type": "string"
                },
                "prescription_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "pharmacy.DispenseRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/pharmacy.DispenseItem"
                    }
                }
            }
        },
        "pharmacy.MedicineResponse": {
            "type": "object",
            "properties": {
                "batch_number": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "formulation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_expired": {
                    "type": "boolean"
                },
                "low_stock": {
                    "type": "boolean"
                },
                "low_stock_threshold": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "near_expiry": {
                    "type": "boolean"
                },
                "stock_qty": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "pharmacy.PrescriptionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "doctor_name": {
                    "type": "string"
                },
                "encounter_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/prescriptions.PrescriptionItemResponse"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "pharmacy.StockAdjustRequest": {
            "type": "object",
            "required": [
                "quantity"
            ],
            "properties": {
                "batch_number": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
        "pharmacy.UpdateMedicineRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "batch_number": {
                    "type": "string",
                    "maxLength": 100
                },
                "expiry_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "formulation": {
                    "type": "string",
                    "maxLength": 100
                },
                "is_active": {
                    "type": "boolean"
                },
                "low_stock_threshold": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "unit": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "prescriptions.PrescriptionItemResponse": {
            "type": "object",
            "properties": {
                "anupana": {
                    "type": "string"
                },
                "dispensed_qty": {
                    "type": "integer"
                },
                "dose": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "formulation": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "medicine": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "route": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}